onto styling canvas index page
