+++ src/components/Canvas/Canvas.scss
@@ -1,0 +1,30 @@
+.toggle-btn {
+    padding: 8px 25px;
+    margin: 10px;
+    background-color: rgb(25, 64, 131);
+    border: none;
+    border-radius: 50px;
+  }
+
+  .canvas-show {
+    height: 100%;
+    width: 100%;
+    padding: 30px;
+  }
+  
+  .canvas-header {
+    text-align: center;
+  }
+  
+  .canvas-container {
+    display: grid;
+    grid-template-columns: 1fr 1fr;
+    max-height: 625px;
+  } 
+  
+  .canvas {
+    width: 600px;
+    margin-top: 10px;
+    text-align: center;
+    box-shadow: 0px 50px 100px 0px rgba(9,24,50, 0.25);
+  }
