--- src/components/presentational/styles/CanvasCard.scss
@@ -1,23 +1,0 @@
-.canvas-card {
-  padding: 1em;
-  background-color: rgb(31,34,41);
-  color: white;
-  border-radius: 10px;
-  // box-shadow: 0 20px 20px rgba(31,34,41,.25);
-  -webkit-box-shadow: 10px 10px 5px 0px rgba(0,0,0,0.75);
-  -moz-box-shadow: 10px 10px 5px 0px rgba(0,0,0,0.75);
-  box-shadow: 10px 10px 5px 0px rgba(0,0,0,0.75);
-  transition: all 0.3s ease 0s;
-
-    h4 {
-      margin: 1em 0;
-    }
-    p {
-      font-size: .75em;
-      margin: 1em 0;
-    }
-  }
-
-.canvas-card:hover {
-  box-shadow: 0 40px 40px rgba(0,0,0,.25);
-}
