--- src/components/styles/NavBar.scss
@@ -1,47 +1,0 @@
-
-header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    padding: 1em;
-    svg {
-        height: 2em;
-    }
-    ul {
-        list-style: none;
-        li {
-            display: inline-block;
-            margin-right: 4em;
-            font-weight: 500;
-            text-decoration: none;
-            a {
-                color: white;
-                text-decoration: none;
-            }
-
-            button {
-                color: white;
-                background: none;
-                border: none;
-                font-size: 1em;
-                font-family: "Futura", sans-serif;
-                cursor: pointer;
-            }
-
-            a, button {
-                transition: all 0.3s ease 0s;
-            }
-            
-            a:hover, button:hover {
-                color: #133e7c
-            }
-        }
-    }
-}
-
-@media (min-width: 800px) {
-
-    
-}
-
-
