+++ src/constants/colors.scss
@@ -1,0 +1,6 @@
+$main-color: #002A3E;
+$main-offset: #012436;
+$accent-one: #01BDC6;
+$accent-two: #FF2079;
+$text: #ffffff;
+$text-offset: rgb(228,228,228);
