+++ src/containers/SongsContainer/SongsContainer.scss
@@ -1,0 +1,12 @@
+.song-container {
+    ul {
+        list-style: none;
+        li {
+            color: black;
+        }
+    
+        li:hover {
+            cursor: pointer;
+        }
+    }
+}
