--- src/containers/styles/CanvasesContainer.scss
@@ -1,33 +1,0 @@
-.canvases-index {
-    display: flex;
-    flex-flow: column;
-    height: 100%;
-    background: rgb(19, 20, 23);
-    padding: 4em 2em;
-    color: white;
-
-    #divider {
-        background-color: rgb(2,189,252);
-        width: 100%;
-        height: .25em;
-    }
-}
-
-.canvas-form {
-    justify-content: center;
-}
-
-.canvas-list {
-    overflow: auto;
-    ul {
-        display: flex;
-        list-style: none;
-        display: flex;
-        flex-wrap: wrap-reverse;
-        justify-content: space-around;
-
-        li {
-            margin: 1em;
-        }
-    }
-}
