--- src/containers/styles/SongsContainer.scss
@@ -1,12 +1,0 @@
-.song-container {
-    ul {
-        list-style: none;
-        li {
-            color: black;
-        }
-    
-        li:hover {
-            cursor: pointer;
-        }
-    }
-}
