+++ src/pages/landing/Landing.scss
@@ -1,4 +1,4 @@
-@import "../../colors.scss";
+@import "../../constants/colors.scss";
 
 main {
     div {
@@ -22,6 +22,7 @@ main {
             text-transform: uppercase;
             font-size: 1em;
             margin-top: 3em;
+            cursor: pointer;
         }
     }
 }
@@ -32,6 +33,7 @@ main {
     right: 0;
     width: 100%;
     height: 100vh;
+    z-index: -1;
     svg {
         bottom: -50px;
         right: -160px;
